+++ changelog.md
@@ -1,5 +1,33 @@
 # What's new in Tmpo
 
+## 2.2.0
+
+### Feature
+
+- Add template helper methods:
+  camelCase, CONSTANT_CASE, kebab-case, lowercaser, PascalCase, snake_case, UPPERCASE
+
+### Fix
+
+- git remote creation
+- deep target directoy
+
+## 2.1.0
+
+### Feature
+
+- Folder name can now be changed with the target-directory flag or input
+
+## 2.0.0
+
+### Feature
+
+- Add the possibility to share single templates via git repository with other users.
+
+### Breaking
+
+- Update format of configuration file.
+
 ## 1.8.1
 
 ### Fix